requirements.txt
biopython==1.78
captum==0.4.1
ConfigSpace==0.4.21
logomaker==0.8
matplotlib==3.5.3
numpy==1.21.5
pandas==1.3.5
PyYAML==5.4.1
ray==1.13.0
scipy==1.7.3
seaborn==0.12.2
torch==1.11.0
torchaudio==0.11.0
torchtext==0.4.0
wandb==0.12.10
tqdm==4.64.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
hpbandster==0.7.4
